requirements.txt
SpeechRecognition==3.8.1
pyttsx3==2.90
spotipy==2.19.0
youtube-search==2.1.0
python-dotenv==0.19.2
Flask==2.0.1
PyAudio==0.2.11
PyQt5==5.15.6
pycaw==20181226
comtypes==1.1.11
pynput==1.7.6
requests==2.28.1
newsapi-python==0.2.6
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.0.2
wolframalpha==5.0.0
face-recognition==1.3.0
opencv-python-headless==4.5.3.56
psutil==5.8.0
GPUtil==1.4.0
schedule==1.1.0
transformers==4.21.0
torch==1.9.0
scikit-learn==0.24.2
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
pytz==2021.3
